Fix behavior in non-interactive environments
gitlike-cli has an issue where it appends an stdin socket to the list of params,
if stdin is not a tty. `glob.replace` crashes if given a socket object.
https://github.com/jedmao/gitlike-cli/issues/3

Such a situation might happen in any non-interactive environment, for example, CI.
The fix ensures that only strings are passed further.